Level Development
New strategy on how levelling works that an offset is created on what level you become. Current code refelects that making both player and enemy object based. Might be a case that in the future an enemy's level could increase if you lose too many battles with an appentice that they become a master